refactor: convert namespace @Helpers to @Utils
