+++ craco.config.js
@@ -8,7 +8,7 @@ module.exports = {
       "@Layouts": path.resolve(__dirname, "src/components/Layouts"),
       "@Services": path.resolve(__dirname, "src/services/"),
       "@Pages": path.resolve(__dirname, "src/pages/"),
-      "@Helpers": path.resolve(__dirname, "src/utils/"),
+      "@Utils": path.resolve(__dirname, "src/utils/"),
       "@Contexts": path.resolve(__dirname, "src/contexts/"),
     },
   },
